+++ src/controllers/slack/ActionControllers.ts
@@ -1,0 +1,7 @@
+export default class ActionControllers {
+  public static buttonAction(payload: object, respond: (body: object) => {}) {
+    console.log();
+    console.log(payload);
+    console.log();
+  }
+}
