+++ src/controllers/slack/EventControllers.ts
@@ -1,0 +1,15 @@
+export default class EventControllers {
+  public static reactionAdded(event: any) {
+    // Verify that that channel has been added for watch
+    console.log();
+    console.log(event);
+    console.log();
+  }
+
+  public static reactionRemoved(event: any) {
+    // Verify that that channel has been added for watch
+    console.log();
+    console.log(event);
+    console.log();
+  }
+}
